pyproject.toml
[project]
name = "better-streamlit-authenticator"
version = "0.4.2"
description = "Updated streamlit-authenticator with better security and usability."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "argon2-cffi>=23.1.0",
    "bcrypt>=4.3.0",
    "captcha>=0.7.1",
    "cryptography>=44.0.3",
    "extra-streamlit-components>=0.1.80",
    "pyjwt>=2.10.1",
    "pyyaml>=6.0.2",
    "streamlit>=1.45.1",
    "streamlit-extras>=0.7.1",
    "watchdog>=6.0.0",
]
